:sparkles:  show homepage link
