+++ packages/core/package.json
@@ -60,6 +60,7 @@
     "read-yaml-file": "^2.1.0",
     "rimraf": "3.0.2",
     "tempy": "^2.0.0",
+    "terminal-link": "^3.0.0",
     "tslib": "^2.3.0",
     "update-notifier": "^5.1.0",
     "write-yaml-file": "^4.2.0"
