+++ packages/core/src/utils/constants.ts
@@ -5,3 +5,4 @@ export const STORE_PATH = path.join(os.homedir(), '.neo-store')
 export const LOCK_FILE = 'neo-lock.yml'
 export const CACHE_DIRNAME = '.cache'
 export const NPM_REGISTRY = 'https://registry.npmjs.org/'
+export const ASSETS_BANNER = 'assets/neo.txt'
