+++ packages/core/src/utils/index.ts
@@ -2,6 +2,7 @@ import { fileURLToPath } from 'url'
 import path, { dirname } from 'path'
 import { findUp } from 'find-up'
 import minimatch from 'minimatch'
+import { readPackageUpSync } from 'read-pkg-up'
 
 const __dirname = dirname(fileURLToPath(import.meta.url))
 
@@ -32,4 +33,9 @@ export async function isOffline() {
       })
     })
   })
+}
+
+export const readPkg = () => {
+  const pkg = readPackageUpSync({ cwd: r() })?.packageJson
+  return pkg
 }
